// ImunizacijaApplication/ImunizacijaAppFront/imunizacija/src/modules/shared/enums/Vakcine.ts
export enum Vakcine {
    PFIZER = 'Pfizer|BionTech|6573849',
    SPUTNIKV = 'Sputnik V|The Gamaleya Center|584307',
    SINOPHARM = 'Sinopharm|China National Pharmaceutical Group Corporation|3648294',
    ASTRAZENECA = 'AstraZeneca|AstraZeneca PLC Company|635478',
    MODERNA = 'Moderna|Technology Square Cambridge, Massachusetts|2351223'
}